Add files via upload
The goal of this notebook is to analyze and explore a dataset on airlines customer satisfaction. By leveraging data science techniques, we aim to uncover insights and patterns that can shed light on the key drivers of customer satisfaction in the airline industry. Through this analysis, we aim to provide valuable findings and actionable recommendations that can assist airlines in enhancing their services and delivering exceptional customer experiences.

Specifically, this notebook focuses on examining various aspects of the airline travel experience, including seat comfort, inflight services, online support, and more. By analyzing these factors, we identified which aspects have the most significant impact on customer satisfaction. Furthermore, we explored any relationships or correlations between different variables to gain a comprehensive understanding of the factors that contribute to overall satisfaction.

Through this exploration, our goal is to not only provide insights into customer satisfaction but also offer practical recommendations to airlines. These recommendations can help them prioritize improvements, allocate resources effectively, and ultimately create a more satisfying and enjoyable travel experience for their passengers.

By sharing these findings and recommendations, we hope to contribute to the ongoing efforts of the airline industry in optimizing customer satisfaction and driving positive customer experiences.